+++ ecl_linear_algebra/include/ecl/linear_algebra/sophus.hpp
@@ -1,0 +1,19 @@
+/**
+ * @file /ecl_linear_algebra/include/ecl/linear_algebra/sophus.hpp
+ */
+/*****************************************************************************
+** Ifdefs
+*****************************************************************************/
+
+#ifndef ecl_linear_algebra_LINEAR_ALGEBRA_SOPHUS_HPP_
+#define ecl_linear_algebra_LINEAR_ALGEBRA_SOPHUS_HPP_
+
+/*****************************************************************************
+** Includes
+*****************************************************************************/
+
+#include "sophus/helpers.hpp"
+#include "sophus/formatters.hpp"
+#include "sophus/interpolators.hpp"
+
+#endif /* ecl_linear_algebra_LINEAR_ALGEBRA_SOPHUS_HPP_ */