+++ ecl_linear_algebra/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 ^^^^^^^^^
 
+Forthcoming
+-----------
+* expose sophus and add some sophus helpers/utilities.
+
 0.61.0 (2014-09-12)
 -------------------
 * eigen matrix formatter api aligned with other ecl float formatters